+++ web/front/src/css/register.css
@@ -14,7 +14,7 @@ body::before {
   width: 100%;
   height: 100%;
   /* background:#cc189f; */
-  background-image: url('./assets/register-bg.jpg');
+  background-image: url('../assets/register-bg.jpg');
   background-position: center;
   background-size: cover;
 }
@@ -74,27 +74,6 @@ h2 {
   transform: translateY(-120%);
 }
 
-.forget {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin: 25px 0 35px 0;
-  color: #fff;
-}
-
-#remember {
-  accent-color: #fff;
-}
-
-.forget label {
-  display: flex;
-  align-items: center;
-}
-
-.forget label p {
-  margin-left: 8px;
-}
-
 .wrapper a {
   color: #efefef;
   text-decoration: none;
@@ -123,7 +102,7 @@ button:hover {
   background: rgba(255, 255, 255, 0.15);
 }
 
-.register {
+.login {
   text-align: center;
   margin-top: 30px;
   color: #fff;
