+++ web/front/src/templates/entrance.jsx
