+++ web/front/src/templates/login.jsx
@@ -1,4 +1,4 @@
-import './login.css'
+import '../css/login.css'
 
 function LogIn() {
   return (
