+++ web/front/src/templates/register.jsx
@@ -1,4 +1,4 @@
-import './register.css'
+import '../css/register.css'
 
 function Register() {
   return (
